+++ tests/cefclient/browser/client_handler.cc
@@ -1424,7 +1424,7 @@ void ClientHandler::SetOfflineState(CefRefPtr<CefBrowser> browser,
 
 void ClientHandler::FilterMenuModel(CefRefPtr<CefMenuModel> model) {
   // Evaluate from the bottom to the top because we'll be removing menu items.
-  for (int i = model->GetCount() - 1; i >= 0; --i) {
+  for (size_t i = model->GetCount() - 1; i >= 0; --i) {
     const auto type = model->GetTypeAt(i);
     if (type == MENUITEMTYPE_SUBMENU) {
       // Filter sub-menu and remove if empty.
