+++ tests/cefclient/browser/views_style.cc
@@ -100,7 +100,7 @@ void ApplyTo(CefRefPtr<CefMenuModel> menu_model) {
                          g_background_hover_color);
 
   // Recursively color sub-menus.
-  for (int i = 0; i < menu_model->GetCount(); ++i) {
+  for (size_t i = 0; i < menu_model->GetCount(); ++i) {
     if (menu_model->GetTypeAt(i) == MENUITEMTYPE_SUBMENU)
       ApplyTo(menu_model->GetSubMenuAt(i));
   }
